ChartLayers.tsx BaseLayers.tsx Removed some unused code from CommandControl.jsx
